Fix empty meta description tags and check params consistently (#1445)
The `{{ else if ne .Summary "" }}` check is always evaluating to true (probably because `Summary` is typed as `template.HTML` instead of a `string` in Hugo's source), so we go down this path, but when then using `.Summary`, the value is empty. This results in many pages having empty `<meta name="description">`  instead of falling back to using the default description. The fix is to check `{{ else if .Summary }}` which will only evaluate to true if `.Summary` is set and is a non-zero value.

While making changes here, I cleaned up how we access other params throughout. We'll often use `if isset .Params "foo"` to check if the `foo` param is set, but this only checks if the param is set -- it doesn't check if the param actually has a value. The fix is to just use `if .Params.foo`, which will only be true if `foo` is set and is not: `false` (boolean), `0` (integer), or a zero-length array, slice, map, or string.