Remove LEARN button from docs homepage
Removed the "LEARN" button on the docs home page as there's no longer a need for it. Previously, it was a way to get to the Reference page to see the expanded Reference TOC, but now that we're expanding by default, there's little need for it. And it never felt quite right having it named "LEARN" but linking to "Reference". Now we have the single CTA we want to "GET STARTED".