Add auto-generated tutorials and indexes (#1620)
This change adds auto-generated tutorials, using our [examples repo](https://github.com/pulumi/examples) as its source. Namely, we scrape the README.mds. We do some on-the-fly rewriting of certain content to make it work on the docs website -- including adding appropriate links back to the GitHub source code -- however, for the most part, the content is taken as-is. There are plenty of examples that need to be tidied up for consistency, to make the titles better, and so on, but this seems like a significant overall improvement so we plan to land roughly as-is.

* Add auto-generated tutorials and indexes

* Update URLs to use new post-ToC update structure

* Add a Featured Tutorials section to AWS

* Add curated "Featured Tutorials" sections

* Remove numbers from tutorial titles

* Add links to the tutorials source code

* Remove "all tutorials" index

* Don't link to AKS from GCP tutorials page

* Add a VIEW CODE button

And also wordsmith some of the "get source code first" note.

* Use linktitle for shorter titles

This ensures that the navigation name and the breadcrumb match.

* Update text for requesting a new tutorial

* Add mktutorial to `make gen`

* Add a nicer tutorials landing page

* Simplify the VIEW CODE button

* Rewrite relative Markdown links

This change automatically rewrites Markdown links that are
relative to our GitHub repo, so that we don't end up with broken
links in the resulting documentation website.

* Fix some broken links

* Ignore some example links

Our examples contain a lot of illustrative links that are either
synthetic or not meant to be publicly accessed. Ignore them during
broken link checking.