Do an editorial pass over programming model docs (#2241)
* Do an editorial pass over programming model docs

This is an editorial pass over the programming model docs, as part
of #1618. This includes rearranging ordering to minimize forward
references, getting rid of anything that smells "TypeScript-only",
restructuring headers so the ToC is a bit more structured,
wordsmithing, fleshing out examples, and trying to simplify the
presentation of content -- while also embellishing where it helps.

This is not done. The Go and TypeScript pages are empty. There is
still quite a number of TypeScript-only hyperlinks and terminology.

* Add Vim swp files to gitignore

* Use shortcodes to cut down on language-specific text/links

This change adds some shortcodes for common language-specific
text and associated links.

For instance, outputs are Output in JS, Output<T> in TS,
Output[T] in Python, Output in Go, and Output<T> in C#, and go
to very different places hyperlink-wise.

Because of the way our language switcher JavaScript works, we
display the JS version by default until it kicks in and picks
up the user's choice. This avoids awkward flashes in which all
or none of the options are available or missing.

I'm not done applying this to all possible codes, but at least
it lays the foundation to start paying down this debt.

* Fix a couple hygiene issues

* Clarify conditional execution of apply

* Fix input shortcode

* Add shortcodes for most language-specific things

This eliminates most (all?) of the language-specific texts and
links in the programming model document, replaced with shortcodes.

* Fix up some loose ends

* Add Go language page and Get Started guides

* Fix linter errors

* Apply suggestions from code review

Co-Authored-By: Christian Nunciato <c@nunciato.org>

* Apply additional code review feedback

* Use spaces, not tabs.

* Use tailwind styles, not inline CSS.

Co-authored-by: Christian Nunciato <christian@pulumi.com>