Update tutorials (#3260)
Re-run mktutorial.sh to regenerate tutorial content from the Examples repo.

It's not immediately clear to me we *want* to keep all the newly added tutorials here, as the examples are of a varying level of quality/completeness.  We could alternatively only accept the "modifications" here and leave out the new tutorials for now.  But broadly speaking the additions help flesh out multi-language coverage for a number of areas.  I'm inclined to just keep everything here for now, until we decide on a different strategy for tutorials overall.