Eliminate "real languages" from our website (#4032)
* Eliminate "real languages" from our website

We know that this terminology is unfriendly and can be seen as
gatekeeping. We've moved away from it over time, but there are many
traces of it in our documentation -- most notably, even on our
homepage. Instead of "real languages", let's use "familiar languages",
and in places where deeper explanation makes sense in context, we
can say "familiar, general-purpose languages."

The current 2nd-level blurb on the homepage isn't very good right now
anyhow, so this is a convenient change. As we make progress on various
marketing efforts, we will likely make more extensive changes, but
for now this at least fixes this problem. We recently A/B tested
updates without much of an impact, so let's move ahead with this change
and then if we want to test alternatives to "familiar", we can do that.
For instance, we considered "popular", etc.

* Also eliminate "real code"