Fix poor Python API docs formatting
It appears we made a change at some point that made it so that our
CSS selector no longer selected the Python API docs as intended, leading
to virtually illegible rendering of those docs. Unless I'm mistaken, we
only use these docs for the Pulumi SDK now-a-days, and not the other
providers, so the fix just targets the styles to that section.

This isn't the first time this has happened (see
https://github.com/pulumi/docs/pull/1397), so I'd love to know if there
are any ideas on what we can do to structurally avoid this breaking again!