Fix the EKS networking example
This example doesn't compile. Unfortunately, the AWSX VPC constructor
appears to require an args bag, even though all properties of it are
optional. Also, the way you pass subnets seems to have changed in the
EKS class -- or maybe this was always a bug -- but the examples I
found in the EKS repo itself suggested this is the right way to do it.

Fixes #325.