docs: fish completion generator example fix
I believe there is a mistake in the docs when generating auto completion
for the fish shell.

It looks like a general code snippet has been used and the `yourprogram`
has not been replaced with `pulumi`.

This did not work until I changed the fish completion filename to
`pulumi.fish` which I have done in the example command shown in the
docs.

Signed-off-by: Jonathan Parton <jon.parton@gmail.com>