Break Resource page into seperate pages for each core concept
Even after breaking the Programming Model content down, this Resources page still contained a *lot* of core content.  This made a lot of important features feel very "hidden", like resource options, components, dynamic providers, and more.

It also led to suboptimal search results, both in Google and in our own internal site search, because searchs for specific concepts (like "pulumi aliases") would return the "Resources" page, which didn't even rank particularly well as it doesn't "feel" like definitive content for this specific topic.

This change breaks this page into 18 sperate pages.  It keeps the "Custom Resouce" content in the main Resources page (since that content is nearly entirely actually not specific to Custom Resources anyway), pulls the Custom Resource, Resource Providers and Dynamic Providers content into separate top level items, and pulls the Resource Options content into it's own index with a separate page per resource option.

Some of the resulting pages are fairly light in content, but this actually offers room to flesh these out more fully now that they have room to breathe.

Reading through this content as part of the refactor, there is quite a bit of cleanup we could still do here.  I took a few liberties to clean up some things as part of this, but mostly focused on the refactor instead of changes to prose.

This also introduces a set of client-side redirects to the resources page, so that each of the previous anchors which now lives elsewhere will get redirected in the client, following the same strategy used in the Programming Model page refactor.