Add PULUMI_TEST_MODE to guides/testing/unit
I was pulling my hair out trying to figure out why `pulumi.isDryRun`
was returning `false` when I was running my unit tests, given that
[the docs on the method][1] say:

> Note that we always consider executions in test mode to be
> “dry-runs”, since we will never actually carry out an update, and
> therefore certain output properties will never be resolved.

I’m still not sure if/where there is any canonical documentation on
“test mode” but I feel that it should probably always be enabled when
running unit tests. I’d prefer that there be a programmatic/imperative
way to enable it, e.g. `runtime.setDryRun(true)` or something similar,
but in lieu of that I feel the examples should include setting the env
var so as to plant seeds of awareness about this feature and how it
works.

[1]: https://www.pulumi.com/docs/reference/pkg/nodejs/pulumi/pulumi/runtime/#isDryRun