guides/testing/unit: show how to set dryRun flag
Which, as per the discussion in pulumi/pulumi-hugo#1642, should
really/probably be renamed to `isPreview` or something along those
lines.

Anyway, it looks like the C# example already includes passing a value
for this flag, and I don’t know enough Go to add it to the Go example.

Now that I’m writing this commit message, the situation with `dryRun`
is confusing enough (especially given the docs on `runtime.isDryRun`)
that it could probably merit adding a small dedicated section to this
page. I’ll ask in the PR whether that’d be desirable.