Include caveat on "simple" and "easy"
While we obviously want every Pulumi customer to have a wonderful experience with no trouble at all, the use of the words "simple" "easy" and their adjectives in our documentation can be alienating when folks run into trouble during a tutorial or Pulumi operation. What is easy to one person may be difficult for the next.