get-started/*/create-project: Add prompt examples
In the Get Started documentation,
when we tell users to Create a New Project with `pulumi new`,
we include a handful of short paragraphs next to each other
with instructions.

This is not very skimmable:

- keywords like project name/description/stack name
  are hidden in the text
- the actual instruction of what to do is also just hidden
  in the text

For someone following along, it's better to have keywords
related to prompts they're seeing highlighted,
and to include a sample output they can follow along to.

This changes the create-project sections
of all four Get Started with Pulumi pages
in the following way:

- project name and description are bold
- change "name of a stack" to "a stack name", and bold "stack name"
- bold out any provider-specific configuration prompts
- add the output of `pulumi new` after each step;
  I've altered the output slightly to not be language-specific

Before this change, only the k8s tutorial included a sample prompt,
but that was at the top of the text rather than interspersed with it.