STYLE-GUIDE: Suggest semantic line breaks
Per Slack discussion,
changes the recommendation on writing paragraphs
to suggest semantic line breaks instead.

The rationale behind the original recommendation was that
it made it easier to suggest changes to a pargraph during code review.
However, that's also possible with multi-line pargraphs
because GitHub supports multi-line comments and suggestions.

This new recommendation includes a new rationale for the choice
as well as a note about leaving existing documentation alone
until it's modified.