shortcode/install-go: Recommend "supported" versions of Go
Some of our documentation currently states that
we support Go 1.18 or higher.
This is not accurate.
We support only those versions of Go that are supported by upstream Go:
typically the latest two minor versions (e.g., 1.20 and 1.19 right now).

This tweaks the wording there to just "supported versions"
instead of hard-coding 1.18 or 1.19
and having to update that 6 months from now.