Hide broken icon in python docs
Sphinx is now emitting a private unicode character ("\F0C1") corresponding to a FontAwesome chain icon next to headers, which renders as an empty box on our site.

Previously, it emitted a "¶" symbol ("\B6"), which we hid with CSS since we have our own way of injecting header links client side with JavaScript.

This change updates the styles to also hide the new unicode character.