archetypes: Use Semantic Line Breaks per Style Guide
We changed the style guide in #2474
to recommend Semantic Line Breaks for new content
while leaving old content as-is.

This had less of an effect
because our generated sample content does not comply with the guidance.

This updates all archetypes (templates for generated files)
to use semantic line breaks in both content and comments,
so that future content is more likely to match the guidance.