concepts/stacks references: explain OutputDetails
This adds a new section to the StackReference section,
explaining how to get values out of a stack reference.

It covers both getOutput and getOutputDetails,
with examples of when one is preferred over the other.

getOuput is primarily suggested when the result
is input to another resource (after some transforms)
and getOutputDetails is suggested when you need direct access.

Code samples for getOutput were adapted from existing cases.
For getOutputDetails, all code samples were manually verified
with a sample project locally.

Notes:

- This functionality does not currently work in Python
  because we do not provide the ability to use `await`
  in Python programs. Ref: pulumi/pulumi#12172
- Although this works in Java, that version of pulumi-java
  has not yet been released.
  We will not land this before releasing pulumi-java.
  Ref: pulumi/pulumi-java#1042

Resolves #2547