Make the migration guide link more noticeable
A user [reported](https://github.com/pulumi/pulumi-aws/pull/2584#issuecomment-1699344037)
that a breaking change was missing from the blogpost, and didn't look at
the migration guide. When I went to point out the migration guide in the blogpost, I
didn't see the link since it looked like it was part of "### MaxItemsOne and
Deprecations".