fix: callback function docs not available
The `CallbackFunction` is a node only function which means that it does
not have any general documentation. The link that is provided to find
more information does not have any information on the
`CallbackFunction`. Since we do not publish any language specific docs,
I don't think there is any better place to point the user to.

fixes https://github.com/pulumi/pulumi-aws/issues/2035